9.1.3 Добавит в `provider` проверку на доступность интернета и зареджектит промисы, если интернета нет
Добавили скелет логики обработки состояния интернета и тудушки. Задача не простая и займет значительное время. Важно не терять мысль и контекст. Скелет и тудушки позволят нам быстрее возвращаться в контекст задачи. Дальше будем заполнять скелет живой логикой.

+++ src/api/provider.js
@@ -1,21 +1,45 @@
+const isOnline = () => {
+  return window.navigator.onLine;
+};
+
 export default class Provider {
   constructor(api) {
     this._api = api;
   }
 
   getTasks() {
-    return this._api.getTasks();
+    if (isOnline()) {
+      return this._api.getTasks();
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   createTask(task) {
-    return this._api.createTask(task);
+    if (isOnline()) {
+      return this._api.createTask(task);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   updateTask(id, task) {
-    return this._api.updateTask(id, task);
+    if (isOnline()) {
+      return this._api.updateTask(id, task);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 
   deleteTask(id) {
-    return this._api.deleteTask(id);
+    if (isOnline()) {
+      return this._api.deleteTask(id);
+    }
+
+    // TODO: Реализовать логику при отсутствии интернета
+    return Promise.reject(`offline logic is not implemented`);
   }
 }
